automated_insight_engine/requirements.txt
Flask>=2.0.0
pandas>=1.5.0
sqlalchemy>=1.4.0
python-dotenv>=0.21.0
openai>=0.27.0
reportlab>=3.6.0
python-pptx>=0.6.21
werkzeug>=2.0.0
